Create utterances.json (#14)
Hi there, we are starting to expand the scope of the IBM Cloud Support Chatbot - and as part of this expansion, we are including multiple IBM doc. Our chatbot currently consumes the FAQ and troubleshooting questions and answers - and uses the training data captured in this utterances file to build out intents. Please can you review this file ( this file also supports the FAQs that Jeanmarie updated in a different PR) and let me know if you have any questions. Once you have merged and approved this file - I will build these intents into Chatbot. Thank you in advance for your help Claire